Start merging blog sources and built blog content
When I started the blog, I didn't find a good way to have blog
sources and the built blog content in one repository. So I used
the repository "blog-source" for the content and this repo
for the built blog content. This means a build & copy & commit
workflow which is not really pleasant. In the meantime I figured
out a way how to do it with the Travis CI. This change here is
the first commit to enable this.

Issue #1