Complete merge of blog source and publishable github pages
The previous commit made all the magic possible which is necessary
to automatically publish github pages every time I commit content
to the "sources" branch. This commit completes it, simply by
adding the build artifacts to a gitignore file. The output of
the _website dir is optional. That's the dir which contains the
build artifacts.

Closes #1