logrotate: finalize the content and pre-date the release
The content should be release-worthy now. To come to a somewhat
regular release interval of stuff, I decided to go for
"the last Friday of the month".
Not sure if can keep that idea up though, but at least it's
something. A follow up change will move the file into the
appropriate folder.