Finalize book recommendation "phoenix project".
I think the structure and content is now good enough to publish.
It's not a "review" anymore -- it never really was -- it is a
recommendation with a lot of words... Which is fine, as there
are enough real book reviews out there, I don't need to add to
those. It's basically me, writing down, what I would otherwise
tell to friends and co-workers.