meta: add description and keywords
Right now Google uses the last entry in "Recent Posts" to
show an excerpt when it shows my page in the search results.
I'd like to have a consistent message there. The last post
changes (hopefully) often.
This change adds HTML meta elements, in the hope that Google
uses this information to create the excerpt.