8: fix warning for "fontawesome_link_cdn"
Unfortunately, copying the fontawesom.min.css into my _static dir
didn't help, as that file references other files itself in a relative
way. So I ignored the ABlog config options and referenced to the
CDN file via URL like I did with my custom.css file, but with a URI.
The real fix is probably in ABlog itself, to ignore the warning
when providing an URL to the option "fontawesome_link_cdn".
When I start pulling from my fork of the little maintained ABlog,
I might incorporate a fix there, but for now I just want to make the
false positive go away.

Closes #8