21: create Makefile
The various commands for the daily work with the blog
are now in the Makefile. A simple shell script with
a big case switch could have done the same, I guess.
My goal was to get rid of the multiple little scripts
and have a reasonable single point of entry.
The existing scripts and their dependencies get removed
and resolved in a follow up patch.

Related issue #21