23: sort spelling wordlist file and test for it
The spelling word list file will grow over time. To make
it easier to comprehend, sort it alphabetically. To avoid
that this file gets unsorted again, add a make target "test"
which checks for that. That test is "voting" in the Travis
CI too.
Over time, the make "test" target will get more checks to
make sure things are consistent.

Closes #23