25: Drop the "Page" button at the top
The "page" button is only useful if there is no other
means to show the TOC of a page. But I show that on
every post. Navigation from "the middle" of a post can
be done by clicking on a section, which brings you back
to the top, which is fine, I think.
This also solves the CSS issue (in Firefox) that the
navbar on top gets too big, when the search gets wrapped
into the "next line" and hides the header of the post.

Closes #25