13: Document my best practices and style guides
I use the template file to document all the things I see
as reasonable to do. The file can then be copied into a
new post and you have all the things you can do with proper
description. It makes it also easy to reference to, in case
there will be guest authors.

Closes #13