Add script to generate random identifiers
If images and downloadable files are not
uniquely names, sphinx will resolve this
by adding numbers to the files when it copies
them into the _images and _download directory.
This is rather ugly, so this change adds a script
which enables to have unique (enough) identifiers
for those resources. There was also a sphinx issue
with parallel builds, where the sphinx name clash
resolution were confused and linked the wrong images
with each other (I don't find the issue anymore),
which means having the unique identifier from the
start avoids that issue too.