finish content of the OSSE 2017 post
The content is done, spell check doesn't complain, the sources
can be built. It's already in the right directory and the publish
date was also already applied. When it goes live like that, I'm OK.
I'll do another round of proof reading though, so maybe another
change will become necessary.