ignore example-app dir during the doc build
From time to time I will have small example applications for demo
purposes and I want to avoid that this application's doc gets
used in the overall blog doc build process and its restrictions
and rules.
This means in the "example-dir", you're much more free to do as
you like or see fit.