Beautify the recent posts list on the landing page
Instead of being a plain bullet list, it's now a list of posts
which looks a bit nicer.
It also uses the last 8 posts now, as this is the last 4 months of
work, which is what I consider as "recent" (and I don't have more
content which I produced regularly).