Pin Ablog to version 0.8.4
Version 0.9.0, released 2018-02-28, seems to have a change which is
not compatible with my current setup. My script to publish, named
`publish-gh-pages.sh`, has a logic which deletes everything,
before publishing the newly built sources. Which led to a situation,
where the deletion was done, although nothing was built. That gets
fixed in another change. This change here simply, sets things back
to the last known good state.