Use my own asciinema sphinx extension
This removes the need of having all casts in the `_static` folder when I
write a post, as the extension will do it. The extension also gives me a
`..asciinema` directive so that I don't have to use `.. raw` anymore.
As the extension deals with copying the asciicasts into the `_website`
outdir, I can keep the asciicasts near the actual posts.

In the future, I might also use this extension to deal with asciicast
versions. I.e. my current casts are version 1, but newer packages of
`asciinema` use version 2 with newline json which is not compatible with
the current player I use. But this is for another commit.