Fix font-awesome with https
After I enforced the github-pages delivery with HTTPS, the HTTP link to
the CDN for the font-awesome CSS didn't work properly anymore and the
permalink icon for example was shown as weird symbol. The problem was
the mixed content, which is shown in Firefox debugger with F12 key.
This commit fixes that by using the https for the CDN.